requirements.txt
# Slack
slack-bolt>=1.18.0
slack-sdk>=3.23.0

# PDF processing
PyPDF2>=3.0.0

# HTTP & API
requests>=2.31.0
urllib3>=2.0.0

# Salesforce
simple-salesforce>=1.12.0

# Snowflake
snowflake-connector-python>=3.0.0

# Google
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
gspread>=5.11.0
google-api-python-client>=2.100.0

# MCP
fastmcp>=0.1.0

# Data processing
pandas>=2.0.0

# Tableau
tableauserverclient>=0.25.0

# Environment
python-dotenv>=1.0.0
